+++ CHANGELOG.md
@@ -37,6 +37,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## Unreleased
 
+### Bug Fixes
+
+* [#271](https://github.com/babylonlabs-io/vigilante/pull/271) fix: support old and new genesis formats
+
 ## v0.22.0
 
 ### Improvements